requirements.txt
google-generativeai>=0.3.0
mcp>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.0.0 